requirements.txt
bedrock-agentcore
strands-agents
requests
mcp
